create package
refactor project structure
